bug fixes
* bug in static view: error that slot not bound actionBar, because the
action bar is attached to the map, when the map is destroyed, bye bye
actionbar. was largely relying on change in streaming to close the
action bar, which seems to be after map destruction -- narrowed the
close, by moving more of its control to playing out of physics/static (
which is good anyway )
* bug: player dialog isnt quite working, it allows the game to move
ahead even though the window is opened: added "turnControl" to handle
much of the stream-empty switching; reworked states ( adding a pause )
as needed.
* bug: player dialog can be dismissed by clicking on overgrey: dismiss
is now a request ( just send an event ), close removes the window;
reworked popup, etc.
* fixed double quip bug ( erroneously flagged for start *and* end
events )
* bug: if the console is open, and i hover outside the console window
then type "look", the cursor comes back. fixed: various issues about
where states goto on modal closed.
* bug: opening the console when the action bar is open keeps us in
SharedPlay, this was because.... we werent handling external closes of
the action bar ( and the shift from static play to static modal is what
closes the bar ). fixed: 1) "normalized" -dismiss and key escape to
.close() the windows instead of just jumping around. 2) added the close
handler to the action bar
* bug: i can click "under" the action bar, most noticeable in static
view when i click on the alien boy, the char is under one of the
buttons, clicking on that button, a "press" occurs in the world
changing the machine, without ever raising the button's event: fixed:
hover doesnt raise events if mouse is hidden.
* bug: zoom not working in static mode. actionButtons had a stray
hide(false) from a previous attempt; removed.
* bug: clicking around in static view doesnt really work well when the
action bar is open -- needed to dismiss the action bar on press.
* bug fix for goto="" key absorb ( and similar ) has to be in deepest
state, repeated if need be, to avoid exiting the lower states.
* bug: facing in static view ( re: "is this working?!" ) it is working,
but the ugly player.linkup() wasnt being called: it was rotating
something on the former screen.
* zooming while static, errors in log, action bar still displayed in
new view: was getting a hide after destroy; patched